TWW: Remove salvage as a base location source
While salvage locations do add a bunch of locations, they also add an
equal number of progression items to the pool, so the number of unfilled
locations for progression items, that are required to goal the seed,
does not increase, so salvage locations are not actually suitable as a
base location source.

Removing salvage as a base location source lowers the overall chance of
salvage locations being enabled, so the individual chance of salvage
locations being enabled has been increased to accommodate.

This does also mean that the other base location sources now have a
higher chance to be present in a seed, but the individual chances of
those location types being enabled has been left untouched.